requirements.txt
asgiref==3.4.1
Django==3.2.6
django-autoslug==1.9.8
django-ckeditor==6.1.0
django-environ==0.7.0
django-js-asset==1.2.2
djangorestframework==3.12.4
Pillow==8.3.1
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
Unidecode==1.2.0
